requirements.txt
pillow
pyrogram
aiohttp
aiofiles
dotenv_python
